#281: Use only one convertAsciiToHexa function
