+++ src/app/services/cipher.provider.lib.spec.ts
@@ -1,8 +1,9 @@
-import { TestBed, inject } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { readJSON } from 'karma-read-json';
 
 import { CipherProvider } from './cipher.provider';
 import { testCases } from '../utils/jasmine-utils';
+import { convertAsciiToHexa } from '../utils/converters';
 
 declare var CipherExtras;
 
@@ -144,15 +145,6 @@ describe('CipherProvider Lib', () => {
     });
   });
 });
-
-function convertAsciiToHexa(str): string {
-  const arr1: string[] = [];
-  for (let n = 0, l = str.length; n < l; n ++) {
-    const hex = Number(str.charCodeAt(n)).toString(16);
-    arr1.push(hex !== '0' ? hex : '00');
-  }
-  return arr1.join('');
-}
 
 function getSeedTestData(inputHashes, seedKeys, actualAddresses) {
   const data = [];
